Refactor MCP-Universe logging to minimize diff vs main
- Restore tests/utils/logger.py to original StructuredEventLogger
  signatures, adding new methods for MCP-Universe support
- Move HumanReadableLogger to tests/benchmarks/mcp_universe/reporting.py
  (matching AppWorld's pattern)
- Fix README install instructions to use UV_GIT_LFS=1 uv pip install
- Remove redundant bfcl and mcpuniverse-eval optional dep groups
